ゲームシーンから開発シーンへ移植 ゲームシーンを純粋なゲーム用のシーンとして管理
