+++ project/Application/scene/SceneFactory.cpp
@@ -1,6 +1,7 @@
 #include "SceneFactory.h"
 
 #include "GameScene.h"
+#include "DevScene.h"
 #include "SceneIds.h"
 #include "TitleScene.h"
 
@@ -13,6 +14,9 @@ std::unique_ptr<BaseScene> SceneFactory::Create(const std::string &sceneId,
   }
   if (sceneId == SceneId::Game) {
     return std::make_unique<GameScene>();
+  }
+  if (sceneId == SceneId::Dev) {
+    return std::make_unique<DevScene>();
   }
   return nullptr;
 }
