+++ project/Application/scene/SceneIds.h
@@ -4,4 +4,5 @@
 namespace SceneId {
 static const std::string Title = "Title";
 static const std::string Game = "Game";
+static const std::string Dev = "Dev";
 } // namespace SceneId
